+++ app/scripts/fhir/R4/lforms-fhir-sdc.js
@@ -796,7 +796,7 @@ var sdcExport = {
           });
         }
 
-        if(answer.scope) {
+        if(answer.score) {
           ext.push({
             "url" : "http://hl7.org/fhir/StructureDefinition/questionnaire-ordinalValue",
             "valueDecimal" : parseFloat(answer.score)